+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 2f39c61 - v11
+| Author: fjrodafo - Fri, 6 Dec 2024 14:13:20 +0000
+| 
 * Commit: 968f24e - v10
 | Author: FJrodafo - Sun, 6 Oct 2024 19:05:52 +0100
 |